added and fixed some routers for both doctor and patient

+++ backend/src/middlewares/deleteLocalFile.js
@@ -1,11 +1,13 @@
 import fs from "fs";
 
-export const deleteLocalImgFile = async (req, res, next) => {
-     if (!req.file) return next();
+const deleteLocalImgFile = async (req, res, next) => {
+     if (!req.file || !req.body.emailId) return next();
 
      await fs.rm(req.file.path, { force: true }, (err) => {
           if (err) console.log(err);
      });
 
      return next();
-}
+};
+
+export default deleteLocalImgFile;
